prune empty dir script shouldn't care if root is an empty dir

+++ scripts/_prune-empty-dirs.sh
@@ -5,12 +5,14 @@
 # Author: Dan Harms <enniomore@icloud.com>
 # Created: Tuesday, June 13, 2017
 # Version: 1.0
-# Modified Time-stamp: <2017-06-21 11:59:05 dan.harms>
+# Modified Time-stamp: <2017-11-16 09:23:12 dan.harms>
 # Modified by: Dan Harms
 # Keywords: utils helper directory
 
 dir=${1:-$(pwd)}
 
-find "$dir" -mindepth 1 -type d -empty -delete
+if [ -d "$dir" ]; then
+    find "$dir" -mindepth 1 -type d -empty -delete
+fi
 
 # code ends here
